Moves BleManager to util
